+++ MediTrackPlatform.API/Devices/Domain/Model/ValueObjects/TemperatureThreshold.cs
@@ -2,11 +2,11 @@ namespace MediTrackPlatform.API.Devices.Domain.Model.ValueObjects;
 
 public record TemperatureThreshold
 {
-    public int Min { get; }
-    public int Max { get; }
+    public double Min { get; }
+    public double Max { get; }
 
     // Constructor para EF Core y creación
-    public TemperatureThreshold(int min, int max)
+    public TemperatureThreshold(double min, double max)
     {
         if (min >= max)
             throw new ArgumentException("El umbral mínimo no puede ser mayor o igual al máximo.");
@@ -19,7 +19,7 @@ public record TemperatureThreshold
     /// Lógica de negocio encapsulada.
     /// Determina si un valor está FUERA del rango seguro.
     /// </summary>
-    public bool IsViolatedBy(int value)
+    public bool IsViolatedBy(double value)
     {
         return value < Min || value > Max;
     }
